fix author mapping when author is embedded in post
